Center item on select
There was an issue with grid wrapping when selecting an item that could
result in the selected item expanding out of view, at least partially.
Here I force the row placement to be whatever the placement is regularly.

Add a helper utils file, contains the gridPos function that is used for
row. Not using the column value, but it needed to be calculated to get
the correct row.