Setup details to be responsive
Show either the one in grid or the absolute one, depending on screen
size. A lot of tweaking to get that to work correctly due to how classes
are swapped.

Several changes that should probably be its own commits also got
included, like not using the let macro in details, but I had to squash a
lot of wip commits to get to a working state.