Enable scroll into view when selecting item
This is sort-of a patch on a odd issue in regards to setting
grid-row-start to an integer value when the item selected is partially
outside the viewport. This fix will scroll the item into view, but I'd
really like to get to the bottom of why that happens, because it's very
inconsistent.