Fix bottom scroll issue
This is somewhat of a compromise, but the issue is when at the bottom of
the page, the `scrollIntoView` call isn't working due to scroll area
dynamically resizing, until the animation is done there is no center
possible to scroll to. So we fire the scroll call again on animation end
to self correct. When not at bottom this will simply maintain position,
but if at the end of page it will correct centering at the end of the
animation.