Fix non visible focus outline on images
Seems that I incorrectly copy pasted focus-visble:outline-hidden from a
different element that actually wants to hide its outline, to an element
that uses the outline. I never noticed because there was an issue with
that class, that got patched, thereby triggering the invalid class in
this cause, causing outlines to be hidden where it shouldn't.

Refs:
https://github.com/tailwindlabs/tailwindcss/blob/main/CHANGELOG.md#4010---2025-03-05
https://github.com/tailwindlabs/tailwindcss/commit/b0aa20c30eb68dc5d4bb636307e4d245d1c781b0